Construct a Reducer

Constructs a reducer that can be used to show different summaries in a plotscaper figure. Specifically, the function creates a list with three values: an initialfn, a reducefn, and a name.

Usage

reducer(initialfn = NULL, reducefn = NULL, name = NULL)

Arguments

initialfn

An JavaScript initializing function

reducefn

A JavaScript reducer function specified

name

A name for the reducer (a string)

Value

A reducer (which is just a list)

Details

reducefn and initialfn should be interpretable as JavaScript functions (either strings or objects constructed with htmlwidgets::JS). Further:

initialfn should take no arguments and just return some value (i.e. a thunk).

reducefn should take two arguments previous and next and return a result of the same type as previous.

Examples

r <- reducer(initialfn = "() => -Infinity",
             reducefn = "(a, b) => Math.max(a, b)",
             name = "max")
set_scene(mtcars) |>
  add_barplot(c("cyl", "mpg"), options = list(reducer = r))