Set up an interactive scene

This function constructs a skeleton of an interactive plotscaper figure. Specifically, it parses the data to JSON, creates an object of class htmlwidget and sends it additional information that is necessary to set up the figure, such as the variable types, the URL of the websocket server (if interactive), etc.. The figure gets rendered when the resulting htmlwidget object gets printed.

Usage

set_scene(
  data = NULL,
  options = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)

Arguments

data

A dataframe that will be converted to JSON (missing values are currently not supported).

options

A list of options

width

Width of the scene

height

Height of the scene

elementId

An id of the element to render the widget in

Value

An object of class htmlwidget

Examples

set_scene(mtcars) |> add_scatterplot(c("wt", "mpg"))